requirements.txt
Flask
requests
python-dateutil
bleach
psycopg2-binary
python-dotenv